rename to isEqual, compare ref in mobile
