+++ packages/webdriverio/src/commands/element.js
@@ -7,7 +7,6 @@ export const custom$$ = require('./element/custom$$').default
 export const custom$ = require('./element/custom$').default
 export const doubleClick = require('./element/doubleClick').default
 export const dragAndDrop = require('./element/dragAndDrop').default
-export const equals = require('./element/equals').default
 export const getAttribute = require('./element/getAttribute').default
 export const getCSSProperty = require('./element/getCSSProperty').default
 export const getHTML = require('./element/getHTML').default
@@ -21,6 +20,7 @@ export const isClickable = require('./element/isClickable').default
 export const isDisplayed = require('./element/isDisplayed').default
 export const isDisplayedInViewport = require('./element/isDisplayedInViewport').default
 export const isEnabled = require('./element/isEnabled').default
+export const isEqual = require('./element/isEqual').default
 export const isExisting = require('./element/isExisting').default
 export const isFocused = require('./element/isFocused').default
 export const isSelected = require('./element/isSelected').default
